requirements.txt
Brotli==1.0.9
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.2
colorama==0.4.4
cssselect2==0.4.1
Flask==2.0.2
fonttools==4.27.1
html5lib==1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
pdfkit==0.6.1
Pillow==8.3.2
pycparser==2.20
pydyf==0.1.1
PyJWT==2.2.0
pyphen==0.11.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
tinycss2==1.1.0
twilio==7.1.0
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2
wkhtmltopdf==0.2
zopfli==0.1.8
